+++ index.ts
@@ -7,4 +7,4 @@ export * from './tick';
 export * from './typechain-types';
 export * from './utils';
 export * as viem from './viem';
-export * from './helpers';
+export * as helper from './helper';
